code correction

+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/Admin/Admin/AdminSanctionPost.cs
@@ -12,9 +12,9 @@
         public string SanctionPostCode { get; set; }
         public string PostType { get; set; }
         public string SanctionedPostDescription { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
-        public Nullable<int> ModifiedBy { get; set; }
+        public Nullable<long> ModifiedBy { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
     }
 }
