+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/GeneralMaster/GeneralRunningNumbers.cs
@@ -20,9 +20,9 @@ namespace Coditech.API.Data
         public long CurrentSequnce { get; set; }
         public bool IsRowLock { get; set; }
         public bool IsActive { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
-        public Nullable<int> ModifiedBy { get; set; }
+        public Nullable<long> ModifiedBy { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
     }
 }
