+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/Gym/GymMemberBodyMeasurement.cs
@@ -10,9 +10,9 @@ namespace Coditech.API.Data
         public int GymMemberDetailId { get; set; }
         public short GymBodyMeasurementTypeId { get; set; }
         public string BodyMeasurementValue { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
-        public Nullable<int> ModifiedBy { get; set; }
+        public Nullable<long> ModifiedBy { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
     }
 }
