+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/MediaManager/MediaConfiguration.cs
@@ -17,9 +17,9 @@ namespace Coditech.API.Data
         public string Custom2 { get; set; }
         public string Custom3 { get; set; }
         public bool IsActive { get; set; } = true;
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
-        public Nullable<int> ModifiedBy { get; set; }
+        public Nullable<long> ModifiedBy { get; set; }
         public Nullable<System.DateTime> ModifiedDate { get; set; }
     }
 }
